python/requirements.txt
cffi==1.15.1
cryptography==40.0.1
greenlet==2.0.2
numpy==1.24.2
pandas==1.5.3
pwinput==1.0.3
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
